+++ subxt/src/frame/contracts.rs
@@ -91,6 +91,7 @@ pub struct CallCall<'a, T: Contracts> {
     /// Address of the contract.
     pub dest: &'a <T as System>::Address,
     /// Value to transfer to the contract.
+    #[codec(compact)]
     pub value: <T as Balances>::Balance,
     /// Gas limit.
     #[codec(compact)]
@@ -115,45 +116,193 @@ pub struct InstantiatedEvent<T: Contracts> {
     pub contract: <T as System>::AccountId,
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+/// Contract execution event.
+///
+/// Emitted upon successful execution of a contract, if any contract events were produced.
+#[derive(Clone, Debug, Eq, PartialEq, Event, Decode)]
+pub struct ContractExecutionEvent<T: Contracts> {
+    /// Caller of the contract.
+    pub caller: <T as System>::AccountId,
+    /// SCALE encoded contract event data.
+    pub data: Vec<u8>,
+}
 
-    subxt_test!({
-        name: test_put_code_and_instantiate,
-        prelude: {
+#[cfg(test)]
+#[cfg(feature = "integration-tests")]
+mod tests {
+    use sp_keyring::AccountKeyring;
+
+    use super::*;
+    use crate::{
+        balances::*,
+        system::*,
+        Client,
+        ClientBuilder,
+        ContractsTemplateRuntime,
+        Error,
+        ExtrinsicSuccess,
+        PairSigner,
+        Signer,
+    };
+    use sp_core::{
+        crypto::AccountId32,
+        sr25519::Pair,
+    };
+    use std::sync::atomic::{
+        AtomicU32,
+        Ordering,
+    };
+
+    static STASH_NONCE: std::sync::atomic::AtomicU32 = AtomicU32::new(0);
+
+    struct TestContext {
+        client: Client<ContractsTemplateRuntime>,
+        signer: PairSigner<ContractsTemplateRuntime, Pair>,
+    }
+
+    impl TestContext {
+        async fn init() -> Self {
+            env_logger::try_init().ok();
+
+            let client = ClientBuilder::<ContractsTemplateRuntime>::new()
+                .build()
+                .await
+                .expect("Error creating client");
+            let mut stash = PairSigner::new(AccountKeyring::Alice.pair());
+            let nonce = client
+                .account(&stash.account_id(), None)
+                .await
+                .unwrap()
+                .nonce;
+            let local_nonce = STASH_NONCE.fetch_add(1, Ordering::SeqCst);
+
+            stash.set_nonce(nonce + local_nonce);
+
+            let signer = Self::generate_account(&client, &mut stash).await;
+
+            TestContext { client, signer }
+        }
+
+        /// generate a new keypair for an account, and fund it so it can perform smart contract operations
+        async fn generate_account(
+            client: &Client<ContractsTemplateRuntime>,
+            stash: &mut PairSigner<ContractsTemplateRuntime, Pair>,
+        ) -> PairSigner<ContractsTemplateRuntime, Pair> {
+            use sp_core::Pair as _;
+            let new_account = Pair::generate().0;
+            let new_account_id: AccountId32 = new_account.public().into();
+            // fund the account
+            let endowment = 200_000_000_000_000;
+            let _ = client
+                .transfer_and_watch(stash, &new_account_id, endowment)
+                .await
+                .expect("New account balance transfer failed");
+            stash.increment_nonce();
+            PairSigner::new(new_account)
+        }
+
+        async fn put_code(
+            &self,
+        ) -> Result<CodeStoredEvent<ContractsTemplateRuntime>, Error> {
             const CONTRACT: &str = r#"
-(module
-    (func (export "call"))
-    (func (export "deploy"))
-)
-"#;
-            let wasm = wat::parse_str(CONTRACT).expect("invalid wat");
-            let code_hash;
-        },
-        step: {
-            call: PutCodeCall {
-                _runtime: PhantomData,
-                code: &wasm,
-            },
-            event: CodeStoredEvent {
-                code_hash: {
-                    code_hash = event.code_hash.clone();
-                    event.code_hash.clone()
-                },
-            },
-        },
-        step: {
-            call: InstantiateCall {
-                endowment: 100_000_000_000_000,
-                gas_limit: 500_000_000,
-                code_hash: &code_hash,
-                data: &[],
-            },
-            event: InstantiatedEvent {
-                caller: alice.clone(),
-                contract: event.contract.clone(),
-            },
-        },
-    });
+                (module
+                    (func (export "call"))
+                    (func (export "deploy"))
+                )
+            "#;
+            let code = wabt::wat2wasm(CONTRACT).expect("invalid wabt");
+
+            let result = self.client.put_code_and_watch(&self.signer, &code).await?;
+            let code_stored = result.code_stored()?.ok_or_else(|| {
+                Error::Other("Failed to find a CodeStored event".into())
+            })?;
+            log::info!("Code hash: {:?}", code_stored.code_hash);
+            Ok(code_stored)
+        }
+
+        async fn instantiate(
+            &self,
+            code_hash: &<ContractsTemplateRuntime as System>::Hash,
+            data: &[u8],
+        ) -> Result<InstantiatedEvent<ContractsTemplateRuntime>, Error> {
+            // call instantiate extrinsic
+            let result = self
+                .client
+                .instantiate_and_watch(
+                    &self.signer,
+                    100_000_000_000_000, // endowment
+                    500_000_000,         // gas_limit
+                    code_hash,
+                    data,
+                )
+                .await?;
+
+            log::info!("Instantiate result: {:?}", result);
+            let instantiated = result.instantiated()?.ok_or_else(|| {
+                Error::Other("Failed to find a Instantiated event".into())
+            })?;
+
+            Ok(instantiated)
+        }
+
+        async fn call(
+            &self,
+            contract: &<ContractsTemplateRuntime as System>::Address,
+            input_data: &[u8],
+        ) -> Result<ExtrinsicSuccess<ContractsTemplateRuntime>, Error> {
+            let result = self
+                .client
+                .call_and_watch(
+                    &self.signer,
+                    contract,
+                    0,           // value
+                    500_000_000, // gas_limit
+                    input_data,
+                )
+                .await?;
+            log::info!("Call result: {:?}", result);
+            Ok(result)
+        }
+    }
+
+    #[async_std::test]
+    async fn tx_put_code() {
+        let ctx = TestContext::init().await;
+        let code_stored = ctx.put_code().await;
+
+        assert!(
+            code_stored.is_ok(),
+            format!(
+                "Error calling put_code and receiving CodeStored Event: {:?}",
+                code_stored
+            )
+        );
+    }
+
+    #[async_std::test]
+    async fn tx_instantiate() {
+        let ctx = TestContext::init().await;
+        let code_stored = ctx.put_code().await.unwrap();
+
+        let instantiated = ctx.instantiate(&code_stored.code_hash, &[]).await;
+
+        assert!(
+            instantiated.is_ok(),
+            format!("Error instantiating contract: {:?}", instantiated)
+        );
+    }
+
+    #[async_std::test]
+    async fn tx_call() {
+        let ctx = TestContext::init().await;
+        let code_stored = ctx.put_code().await.unwrap();
+
+        let instantiated = ctx.instantiate(&code_stored.code_hash, &[]).await.unwrap();
+        let executed = ctx.call(&instantiated.contract, &[]).await;
+
+        assert!(
+            executed.is_ok(),
+            format!("Error calling contract: {:?}", executed)
+        );
+    }
 }