+++ subxt/test-node/src/command.rs
@@ -16,9 +16,11 @@
 
 use crate::{
     chain_spec,
-    cli::Cli,
+    cli::{
+        Cli,
+        Subcommand,
+    },
     service,
-    service::new_full_params,
 };
 use sc_cli::{
     ChainSpec,
@@ -26,7 +28,6 @@ use sc_cli::{
     RuntimeVersion,
     SubstrateCli,
 };
-use sc_service::ServiceParams;
 
 impl SubstrateCli for Cli {
     fn impl_name() -> String {
@@ -75,21 +76,13 @@ pub fn run() -> sc_cli::Result<()> {
     let cli = Cli::from_args();
 
     match &cli.subcommand {
-        Some(subcommand) => {
-            let runner = cli.create_runner(subcommand)?;
-            runner.run_subcommand(subcommand, |config| {
-                let (
-                    ServiceParams {
-                        client,
-                        backend,
-                        task_manager,
-                        import_queue,
-                        ..
-                    },
-                    ..,
-                ) = new_full_params(config)?;
-                Ok((client, backend, import_queue, task_manager))
-            })
+        Some(Subcommand::BuildSpec(cmd)) => {
+            let runner = cli.create_runner(cmd)?;
+            runner.sync_run(|config| cmd.run(config.chain_spec, config.network))
+        }
+        Some(Subcommand::PurgeChain(cmd)) => {
+            let runner = cli.create_runner(cmd)?;
+            runner.sync_run(|config| cmd.run(config.database))
         }
         None => {
             let runner = cli.create_runner(&cli.run)?;